# internal/api/task_handler.go
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"math"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
	"go.opentelemetry.io/otel"
	"go.opentelemetry.io/otel/propagation"
	"go.opentelemetry.io/otel/trace"

	"github.com/qisanfen666/agentflow/internal/observability"
	"github.com/qisanfen666/agentflow/model"
)

// registerTaskRoutes 注册 Task 域 4 端点（合同：openapi.yaml /api/v1/tasks*）。
func registerTaskRoutes(r gin.IRouter, h *handlers) {
	g := r.Group("/api/v1/tasks")
	g.POST("", h.submitTask)
	g.GET("/:id", h.getTask)
	g.DELETE("/:id", h.cancelTask)
	g.GET("/:id/stream", h.streamTask)
	g.POST("/:id/approve", h.approveTask) // M6 审批流（admin）
	g.POST("/:id/reject", h.rejectTask)
}

// idemTTL 幂等 key 的占位窗口：窗口内同 key 视为重复提交。
const idemTTL = 10 * time.Minute

// auditDetail 任务审计事件的公共字段。session_id 高基数，只进审计明细
// （供 ELK 等按会话聚合），不进 Prometheus label。
func auditDetail(task model.Task) map[string]any {
	detail := map[string]any{"agent_id": task.AgentID, "agent_version": task.AgentVersion}
	if task.SessionID != "" {
		detail["session_id"] = task.SessionID
	}
	return detail
}

// captureTraceparent 捕获当前请求 span 的 traceparent 存进任务记录。
// 这是跨队列的因果载体：队列（Redis/内存）不是 HTTP，标准传播器穿不过，
// 把上下文变成数据随任务走，worker 侧以 span Link 挂回。
// tracing 未启用时 SpanContext 无效，返回空串（零开销关闭）。
func captureTraceparent(ctx context.Context) string {
	if !trace.SpanContextFromContext(ctx).IsValid() {
		return ""
	}
	carrier := propagation.MapCarrier{}
	otel.GetTextMapPropagator().Inject(ctx, carrier)
	return carrier.Get("traceparent")
}

// taskInput 提交请求体（openapi TaskInput）。
type taskInput struct {
	AgentID      string         `json:"agent_id"`
	AgentVersion int            `json:"agent_version"` // 0 = 锁定当前最新版本
	SessionID    string         `json:"session_id"`    // 可选：归因分组（审计/span 携带，不进指标 label——高基数）
	Payload      map[string]any `json:"payload"`
	TimeoutSec   int            `json:"timeout_sec"`
}

// submitTask 提交任务：校验 Agent -> 锁定版本 -> 创建 pending -> 入队。
func (h *handlers) submitTask(c *gin.Context) {
	var in taskInput
	if err := c.ShouldBindJSON(&in); err != nil {
		badRequest(c, err.Error())
		return
	}
	if in.AgentID == "" {
		badRequest(c, "agent_id is required")
		return
	}
	// 治理守门（M6）：限流/预算等规则链在入队前评估，拒绝统一 429 + 治理码
	if v := h.deps.Policy.Evaluate(c.Request.Context(), model.Task{
		AgentID: in.AgentID, SessionID: in.SessionID, Payload: in.Payload,
	}); !v.Allowed {
		if v.RetryAfter > 0 {
			c.Header("Retry-After", strconv.Itoa(int(math.Ceil(v.RetryAfter.Seconds()))))
		}
		c.JSON(http.StatusTooManyRequests, gin.H{"code": v.Code, "message": v.Message})
		return
	}

	// 版本锁定：解析出确切的 (agent_id, version) 并固化进任务
	var spec model.AgentSpec
	var err error
	if in.AgentVersion == 0 {
		spec, err = h.deps.Agents.Get(c.Request.Context(), in.AgentID)
	} else {
		spec, err = h.deps.Agents.GetVersion(c.Request.Context(), in.AgentID, in.AgentVersion)
	}
	if err != nil {
		respondErr(c, err) // 404
		return
	}
	if spec.Deleted() {
		c.JSON(http.StatusNotFound, gin.H{"code": codeNotFound, "message": "agent is deleted"})
		return
	}

	status := model.TaskPending
	if spec.RequireApproval {
		status = model.TaskPendingApproval // 高危 Agent：提交即待审，不入队
	}
	task, err := h.deps.Tasks.Create(c.Request.Context(), model.Task{
		Status:       status,
		AgentID:      spec.ID,
		AgentVersion: spec.Version, // 锁定：后续 Agent 更新不影响本任务
		SessionID:    in.SessionID, // 可选归因分组，空则无会话
		TraceContext: captureTraceparent(c.Request.Context()),
		Payload:      in.Payload,
		TimeoutSec:   in.TimeoutSec,
	})
	if err != nil {
		respondErr(c, err)
		return
	}

	// 幂等（task-state.md 第 4 节）：原子占位，同 key 的重复提交返回原任务。
	// 输家手里刚创建的任务成了孤儿（pending 且永不入队）——迁移 cancelled 收尾，
	// 不留永远 pending 的僵尸。
	if key := c.GetHeader("Idempotency-Key"); key != "" {
		existing, inserted, err := h.deps.Idem.PutIfAbsent(c.Request.Context(), key, task.ID, idemTTL)
		if err != nil {
			respondErr(c, err)
			return
		}
		if !inserted {
			_ = task.Transition(model.TaskCancelled)
			_ = h.deps.Tasks.Save(c.Request.Context(), task)
			if orig, gerr := h.deps.Tasks.Get(c.Request.Context(), existing); gerr == nil {
				c.JSON(http.StatusAccepted, orig)
				return
			}
			c.JSON(http.StatusAccepted, task) // 原任务读不到（极端情况）：退回自身
			return
		}
	}

	// M2：提交即入队；执行由 engine.Worker 异步消费（队列化代价：
	// 提交后短暂 pending，毫秒级）。审批流例外：待审任务留在存储里，
	// approve 放行后才入队（队列闸门语义）。
	if task.Status == model.TaskPending {
		if err := h.deps.Queue.Enqueue(c.Request.Context(), task); err != nil {
			respondErr(c, err)
			return
		}
	}
	detail := auditDetail(task)
	if task.Status == model.TaskPendingApproval {
		detail["requires_approval"] = true
	}
	h.audit(c, observability.ActionTaskSubmitted, observability.EntityTask, task.ID, detail)
	c.JSON(http.StatusAccepted, task)
}

func (h *handlers) getTask(c *gin.Context) {
	task, err := h.deps.Tasks.Get(c.Request.Context(), c.Param("id"))
	if err != nil {
		respondErr(c, err)
		return
	}
	c.JSON(http.StatusOK, task)
}

// cancelTask 显式取消：迁移状态 -> 持久化 -> 通知分发器取消 ctx（尽力传播）。
func (h *handlers) cancelTask(c *gin.Context) {
	task, err := h.deps.Tasks.Get(c.Request.Context(), c.Param("id"))
	if err != nil {
		respondErr(c, err)
		return
	}
	if task.Status.Final() {
		// 终态不可迁移：不是错误重试，而是明确告知冲突
		c.JSON(http.StatusConflict, gin.H{"code": codeConflict, "message": "task already in final state " + string(task.Status)})
		return
	}
	if err := task.Transition(model.TaskCancelled); err != nil {
		respondErr(c, err)
		return
	}
	if err := h.deps.Tasks.Save(c.Request.Context(), task); err != nil {
		respondErr(c, err)
		return
	}
	h.deps.Dispatcher.Cancel(task.ID)
	h.audit(c, observability.ActionTaskCancelled, observability.EntityTask, task.ID, nil)
	c.JSON(http.StatusAccepted, task)
}

// approveTask 审批通过：pending_approval -> pending 并入队（唯一的放行动作）。
// Save 先于 Enqueue：落库成功而入队失败是"可见的僵尸"（可查询、可重审），
// 反之会出现"队列里有但状态仍待审"的静默错配。
func (h *handlers) approveTask(c *gin.Context) {
	task, ok := h.loadForReview(c)
	if !ok {
		return
	}
	if err := task.Transition(model.TaskPending); err != nil {
		respondErr(c, err)
		return
	}
	if err := h.deps.Tasks.Save(c.Request.Context(), task); err != nil {
		respondErr(c, err)
		return
	}
	if err := h.deps.Queue.Enqueue(c.Request.Context(), task); err != nil {
		respondErr(c, err)
		return
	}
	h.audit(c, observability.ActionTaskApproved, observability.EntityTask, task.ID, nil)
	c.JSON(http.StatusAccepted, task)
}

// rejectTask 审批驳回：pending_approval -> cancelled（终态）。
func (h *handlers) rejectTask(c *gin.Context) {
	task, ok := h.loadForReview(c)
	if !ok {
		return
	}
	if err := task.Transition(model.TaskCancelled); err != nil {
		respondErr(c, err)
		return
	}
	if err := h.deps.Tasks.Save(c.Request.Context(), task); err != nil {
		respondErr(c, err)
		return
	}
	h.audit(c, observability.ActionTaskRejected, observability.EntityTask, task.ID, nil)
	c.JSON(http.StatusAccepted, task)
}

// loadForReview 审批端点的公共前置：读任务；非待审态直接回 409（终态/已放行不可再审）。
func (h *handlers) loadForReview(c *gin.Context) (model.Task, bool) {
	task, err := h.deps.Tasks.Get(c.Request.Context(), c.Param("id"))
	if err != nil {
		respondErr(c, err)
		return model.Task{}, false
	}
	if task.Status != model.TaskPendingApproval {
		c.JSON(http.StatusConflict, gin.H{"code": codeConflict, "message": "task not awaiting review"})
		return model.Task{}, false
	}
	return task, true
}

// streamTask SSE 透传：把 Hub 中该任务的事件流（含历史回放）按协议原样发给客户端。
// 订阅循环模式：Since 游标 -> 写事件 -> 未结束则等 Notify 唤醒 -> 客户端断开则退出。
func (h *handlers) streamTask(c *gin.Context) {
	taskID := c.Param("id")
	if _, err := h.deps.Tasks.Get(c.Request.Context(), taskID); err != nil {
		respondErr(c, err)
		return
	}

	c.Header("Content-Type", "text/event-stream")
	c.Header("Cache-Control", "no-cache")
	// 告知反向代理禁用缓冲，保证 token 级实时性
	c.Header("X-Accel-Buffering", "no")
	w := c.Writer

	notify := h.deps.Hub.Notify(taskID) // 先订阅再取快照，避免错过事件
	ctx := c.Request.Context()
	i := 0
	for {
		evs, finished, ok := h.deps.Hub.Since(taskID, i)
		if !ok {
			// 任务存在但尚未发布任何事件（刚提交的瞬间），等唤醒
			select {
			case <-notify:
				continue
			case <-ctx.Done():
				return
			}
		}
		for _, ev := range evs {
			data, err := json.Marshal(ev)
			if err != nil {
				continue // 不可序列化事件直接跳过，不打断流
			}
			fmt.Fprintf(w, "data: %s\n\n", data)
		}
		w.Flush()
		i += len(evs)

		if finished {
			fmt.Fprint(w, "data: [DONE]\n\n")
			w.Flush()
			return
		}

		select {
		case <-notify:
		case <-ctx.Done():
			return // 客户端断开，停止透传（任务继续执行不受影响）
		}
	}
}